requirements.txt
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
bidict==0.23.1
biosppy==2.2.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.13
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
durationpy==0.10
email_validator==2.2.0
et_xmlfile==2.0.0
etelemetry==0.3.1
fastapi==0.115.13
fastapi-cli==0.0.7
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.4
frontend==0.0.3
frozenlist==1.7.0
fsspec==2025.5.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
groq==0.33.0
grpcio==1.73.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==33.1.0
looseversion==1.3.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mmh3==5.1.0
mock==5.2.0
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
neurokit2==0.2.12
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
oauthlib==3.3.0
onnxruntime==1.22.0
openai==1.97.1
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pathlib==1.0.1
pillow==11.2.1
posthog==5.3.0
propcache==0.3.2
protobuf==5.29.5
prov==2.1.1
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.1
PyJWT==2.9.0
PyMuPDF==1.26.3
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyWavelets==1.8.0
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==4.1.0
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
sqlparse==0.5.3
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
tools==1.0.2
torch==2.8.0+cpu
tqdm==4.67.1
traits==7.0.2
transformers==4.52.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.3
watchdog==6.0.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wfdb==4.3.0
yarl==1.20.1
zipp==3.23.0
